+++ lib/get_field.dart
@@ -6,13 +6,7 @@ Object getField(Object object, String field) {
   try {
     return reflect(object).getField(Symbol(field)).reflectee;
   } catch (e) {
-    // TODO: добавить: текст ошибки = add: error message
+    // TODO: improve error message
     throw TemplateRuntimeError('$e');
-  }
-}
-
-extension on Map {
-  Object items() {
-    return entries.map((entry) => <Object>[entry.key, entry.value]);
   }
 }